Move abstract input handlers to separate file
